+++ cmd/searcher/main.go
@@ -55,6 +55,9 @@ func main() {
 			FetchTar: func(ctx context.Context, repo gitserver.Repo, commit api.CommitID) (io.ReadCloser, error) {
 				return gitserver.DefaultClient.Archive(ctx, repo, gitserver.ArchiveOptions{Treeish: string(commit), Format: "tar"})
 			},
+			FilterTar: func(ctx context.Context, repo gitserver.Repo, commit api.CommitID) (store.FilterFunc, error) {
+				return search.NewFilter(ctx, repo, commit)
+			},
 			Path:              filepath.Join(cacheDir, "searcher-archives"),
 			MaxCacheSizeBytes: cacheSizeBytes,
 		},
